+++ src/main/java/gaji/service/domain/myRepeat/RepeatException.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.myRepeat;
 
-import gaji.service.domain.room.Event;
+import gaji.service.domain.room.entity.Event;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
