+++ src/main/java/gaji/service/domain/recruite/RecruitPost.java
@@ -2,9 +2,8 @@ package gaji.service.domain.recruite;
 
 import gaji.service.domain.User;
 import gaji.service.domain.common.entity.BaseEntity;
-import gaji.service.domain.common.entity.SelectHashtag;
 import gaji.service.domain.enums.RecruitePostTypeEnum;
-import gaji.service.domain.room.Room;
+import gaji.service.domain.room.entity.Room;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
