+++ src/main/java/gaji/service/domain/room/entity/Event.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.room;
+package gaji.service.domain.room.entity;
 
 import gaji.service.domain.User;
 import gaji.service.domain.myRepeat.MyRepeat;
