+++ src/main/java/gaji/service/domain/room/entity/VoiceChat.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.room;
+package gaji.service.domain.room.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
