+++ src/main/java/gaji/service/domain/roomPost/RoomBoard.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.roomPost;
 
-import gaji.service.domain.room.Room;
+import gaji.service.domain.room.entity.Room;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
