+++ src/main/java/gaji/service/domain/studyMate/Assignment.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.studyMate;
 
-import gaji.service.domain.room.Room;
+import gaji.service.domain.room.entity.Room;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
