+++ src/main/java/gaji/service/domain/studyMate/StudyApplicant.java
@@ -2,7 +2,7 @@ package gaji.service.domain.studyMate;
 
 import gaji.service.domain.User;
 import gaji.service.domain.enums.ApplicantStatus;
-import gaji.service.domain.room.Room;
+import gaji.service.domain.room.entity.Room;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
